+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- The order of list items should be: Critical/Fixes, New, Update, Remove, Underpinnings -->
 <!-- ## [UNRELEASED](https://github.com/roydukkey/chimera-theme/compare/v0.12.0...master) -->
 
+## [UNRELEASED](https://github.com/roydukkey/chimera-theme/compare/v0.12.0...master)
+
+* Add `editor.foldBackground`
+
 ## [0.12.0](https://github.com/roydukkey/chimera-theme/compare/v0.12.0...master)
 
 * Add strike through for markdown following Dark ([@vscode/#43504](https://github.com/microsoft/vscode/issues/43504))